Refactor SleeveDetectionPipeline
