+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/SleevePosition.java
@@ -1,0 +1,10 @@
+package org.firstinspires.ftc.teamcode;
+
+/*
+ * An enum to define the sleeve position
+ */
+public enum SleevePosition {
+    LEFT,
+    CENTER,
+    RIGHT
+}
